+++ static/js/renderer.js
@@ -803,7 +803,6 @@ const app = Vue.createApp({
     await this.probeDocker();
     this.checkMobile();
     this.loadSherpaStatus();
-    this.loadMossStatus();
     this.minilmModelStatus();
     window.addEventListener('resize', this.handleResize);